[FIX] Declaring volume for bundles causes container to not update
This is an extension of the previous commit. 'extra-addons' should be considered in the same way as the 'extra-addons-bundles' folder. When deploying a new version (eg to devdocker) the Odoo service needs to be downed (not just restarted) to have new code changes applied. I believe this is caused by the volume for extra-addons-bundles being declared as a volume in the Dockerfile.